// src/utils/api/api.ts
export interface Faction {
    /**
     * 
     * @type {FactionSymbol}
     * @memberof Faction
     */
    'symbol': FactionSymbol;
    /**
     * Name of the faction.
     * @type {string}
     * @memberof Faction
     */
    'name': string;
    /**
     * Description of the faction.
     * @type {string}
     * @memberof Faction
     */
    'description': string;
    /**
     * The waypoint in which the faction\'s HQ is located in.
     * @type {string}
     * @memberof Faction
     */
    'headquarters': string;
    /**
     * List of traits that define this faction.
     * @type {Array<FactionTrait>}
     * @memberof Faction
     */
    'traits': Array<FactionTrait>;
    /**
     * Whether or not the faction is currently recruiting new agents.
     * @type {boolean}
     * @memberof Faction
     */
    'isRecruiting': boolean;
}


/**
 * The symbol of the faction.
 * @export
 * @enum {string}
 */

export const FactionSymbol = {
    Cosmic: 'COSMIC',
    Void: 'VOID',
    Galactic: 'GALACTIC',
    Quantum: 'QUANTUM',
    Dominion: 'DOMINION',
    Astro: 'ASTRO',
    Corsairs: 'CORSAIRS',
    Obsidian: 'OBSIDIAN',
    Aegis: 'AEGIS',
    United: 'UNITED',
    Solitary: 'SOLITARY',
    Cobalt: 'COBALT',
    Omega: 'OMEGA',
    Echo: 'ECHO',
    Lords: 'LORDS',
    Cult: 'CULT',
    Ancients: 'ANCIENTS',
    Shadow: 'SHADOW',
    Ethereal: 'ETHEREAL'
} as const;

export type FactionSymbol = typeof FactionSymbol[keyof typeof FactionSymbol];


/**
 * 
 * @export
 * @interface FactionTrait
 */
export interface FactionTrait {
    /**
     * 
     * @type {FactionTraitSymbol}
     * @memberof FactionTrait
     */
    'symbol': FactionTraitSymbol;
    /**
     * The name of the trait.
     * @type {string}
     * @memberof FactionTrait
     */
    'name': string;
    /**
     * A description of the trait.
     * @type {string}
     * @memberof FactionTrait
     */
    'description': string;
}


/**
 * The unique identifier of the trait.
 * @export
 * @enum {string}
 */

export const FactionTraitSymbol = {
    Bureaucratic: 'BUREAUCRATIC',
    Secretive: 'SECRETIVE',
    Capitalistic: 'CAPITALISTIC',
    Industrious: 'INDUSTRIOUS',
    Peaceful: 'PEACEFUL',
    Distrustful: 'DISTRUSTFUL',
    Welcoming: 'WELCOMING',
    Smugglers: 'SMUGGLERS',
    Scavengers: 'SCAVENGERS',
    Rebellious: 'REBELLIOUS',
    Exiles: 'EXILES',
    Pirates: 'PIRATES',
    Raiders: 'RAIDERS',
    Clan: 'CLAN',
    Guild: 'GUILD',
    Dominion: 'DOMINION',
    Fringe: 'FRINGE',
    Forsaken: 'FORSAKEN',
    Isolated: 'ISOLATED',
    Localized: 'LOCALIZED',
    Established: 'ESTABLISHED',
    Notable: 'NOTABLE',
    Dominant: 'DOMINANT',
    Inescapable: 'INESCAPABLE',
    Innovative: 'INNOVATIVE',
    Bold: 'BOLD',
    Visionary: 'VISIONARY',
    Curious: 'CURIOUS',
    Daring: 'DARING',
    Exploratory: 'EXPLORATORY',
    Resourceful: 'RESOURCEFUL',
    Flexible: 'FLEXIBLE',
    Cooperative: 'COOPERATIVE',
    United: 'UNITED',
    Strategic: 'STRATEGIC',
    Intelligent: 'INTELLIGENT',
    ResearchFocused: 'RESEARCH_FOCUSED',
    Collaborative: 'COLLABORATIVE',
    Progressive: 'PROGRESSIVE',
    Militaristic: 'MILITARISTIC',
    TechnologicallyAdvanced: 'TECHNOLOGICALLY_ADVANCED',
    Aggressive: 'AGGRESSIVE',
    Imperialistic: 'IMPERIALISTIC',
    TreasureHunters: 'TREASURE_HUNTERS',
    Dexterous: 'DEXTEROUS',
    Unpredictable: 'UNPREDICTABLE',
    Brutal: 'BRUTAL',
    Fleeting: 'FLEETING',
    Adaptable: 'ADAPTABLE',
    SelfSufficient: 'SELF_SUFFICIENT',
    Defensive: 'DEFENSIVE',
    Proud: 'PROUD',
    Diverse: 'DIVERSE',
    Independent: 'INDEPENDENT',
    SelfInterested: 'SELF_INTERESTED',
    Fragmented: 'FRAGMENTED',
    Commercial: 'COMMERCIAL',
    FreeMarkets: 'FREE_MARKETS',
    Entrepreneurial: 'ENTREPRENEURIAL'
} as const;

export type FactionTraitSymbol = typeof FactionTraitSymbol[keyof typeof FactionTraitSymbol];